requirements.txt
torch>=1.11.0+cu102
torchvision>=0.12.0+cu102
Pillow>=9.1.1
pytest>-6.2.*
python-dotenv>=0.19.0
matplotlib>=3.5.2
